Add an explicit extension manifest key
This will be used whenever the extension is unpacked (i.e. all the time,
since we always inject it from disk). Having a known extension id allows
us to keep using CORS in Mockttp to decide who can manage the API - in
this case, only this specific extension, and no others.

Probably not a huge deal, since you're giving a huge amount of trust to
browser extensions in general anyway, and with new Mockttp they can't
admin anything without the session UUID in the URL (which this extension
receives via injected config) but better to be safe.