Fix the webextension key
It's not actually supposed to be the id - it's supposed to be a public
key. That's generable only by packaging the extension (from extensions
page when unpacked) then installing (drag it onto chrome extensions
page) then looking at the manifest in
/home/$user/.config/google-chrome/Default/Extensions/$id/$version.

This then fixes the id in turn, which seems to be a hash of this key.

Private key doesn't actually matter - all installs are transitory, we're
not publishing to the Chrome store, there will be no signed updates.